+++ CHANGES.rst
@@ -4,8 +4,11 @@ Changelog
 Version 0.9.0
 -------------
 
+- Major structure change to better group similar functionality
 - Namespace now operates more like "dict" on init, and can accept both iterable and kwargs
 - Breaking change: removing Cookie Management in favor of separate library
+- Breaking change: removing sort_by
+- Breaking change: removing namespace.from_dict()
 
 Version 0.8.0
 -------------
