+++ reusables/shared_variables.py
@@ -95,7 +95,7 @@ common_variables = {
     },
 }
 
-sizes = Namespace.from_dict({
+sizes = Namespace({
     "kb": 1024,
     "mb": 1024 * 1024,
     "gb": 1024 * 1024 * 1024,
