+++ reusables/string_manipulation/__init__.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+# -*- coding: UTF-8 -*-
+#
+# Part of the Reusables package.
+#
+# Copyright (c) 2014-2017 - Chris Griffith - MIT License
+from .numbers import *
+from .strings import *
